requirements.txt
# requirements.txt
Flask
flask_cors
Flask-SQLAlchemy
Flask-JWT-Extended
python-dotenv
Werkzeug